move user updates from PresenceUpdateHandler to GuildMemberUpdateHandler
